requirements.txt
# Data ingestion
yfinance>=0.2.12
scrappy>=2.8.0
firecrawl>=0.1.0
pdfplumber>=0.9.0
PyPDF2>=3.0.0
requests>=2.28.2

# Vector store and RAG
pinecone-client>=2.2.1
sentence-transformers>=2.2.2
langchain>=0.0.267
langchain-community>=0.0.10

# Analysis tools
pandas>=2.0.0
numpy>=1.24.3

# Language processing
langgraph>=0.0.10
torch>=2.0.0
transformers>=4.30.2

# Voice processing
openai-whisper>=20231117
sarvam-ai-tts>=0.1.0
elevenlabs>=0.2.24
pydub>=0.25.1
sounddevice>=0.4.6

# Web frameworks
fastapi>=0.95.1
uvicorn>=0.22.0
streamlit>=1.22.0

# Cache and performance
redis>=4.5.5

# Utils
python-dotenv>=1.0.0
loguru>=0.7.0
apscheduler>=3.10.1
pytz>=2023.3

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Development tools
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0

# Deployment
python-multipart>=0.0.6
jsonpickle>=3.0.1
